SimpleStatics

Documentation for SimpleStatics.

SimpleStatics.UnorderedPair

SimpleStatics.add_joint!

SimpleStatics.add_member!

SimpleStatics.constrained_array

SimpleStatics.constrained_dof_array_mapping

SimpleStatics.dof_indices

SimpleStatics.force_vector

SimpleStatics.member_angle

SimpleStatics.member_stiffness_matrix

SimpleStatics.n_dofs

SimpleStatics.original_array

SimpleStatics.set_force!

SimpleStatics.solve_member_stresses

SimpleStatics.terminal_joints

SimpleStatics.UnorderedPair — Type

Create a pair of items whose order is independent in hashes, equality, and comparisons.

SimpleStatics.add_joint! — Method

Add a joint to the setup and return the index referring to that joint. The position should be set in meters.

SimpleStatics.add_member! — Method

Add a member to the setup and return the index referring to that member.

SimpleStatics.constrained_array — Method

Construct a view of an array where constrained indices, indicated by a mapping, are removed. This is useful for solving matrices as zero rows will cause det(M)==0.

SimpleStatics.constrained_dof_array_mapping — Method

Find a mapping m[i] -> j where i is the index of a constrained dof array and j is the corresponding index of the unconstrained array.

SimpleStatics.dof_indices — Method

Get the indices of the x and y rows, respectively, in a square DOF-like matrix (e.g., global stiffness matrix) for a vertex (joint) index.

SimpleStatics.force_vector — Method

Get a vector of the unconstrained forces in the setup.

SimpleStatics.member_angle — Method

Get the angle, in radians, the member makes with the global coordinate system's positive X axis.

SimpleStatics.member_stiffness_matrix — Method

Get the global stiffness matrix for a StaticSetup in Newtons / Meter.

SimpleStatics.n_dofs — Method

Naive number of dergees of freedom in the system (i.e., before constraints).

SimpleStatics.original_array — Method

Recover the original array where DOF indices will match with the StaticSetup.

SimpleStatics.set_force! — Method

Set the force at a joint, in Newtons.

SimpleStatics.solve_member_stresses — Method

Negative -> Compressive, Positive -> Tensile, index matches member index

SimpleStatics.terminal_joints — Method

Get the joints IDs that a member connects, found by its member ID.