SimpleStatics.AnchorConstraint

SimpleStatics.NoConstraint

SimpleStatics.UnorderedPair

SimpleStatics.XRollerConstraint

SimpleStatics.YRollerConstraint

SimpleStatics.add_joint!

SimpleStatics.add_member!

SimpleStatics.constrained_array

SimpleStatics.constrained_dof_array_mapping

SimpleStatics.dof_indices

SimpleStatics.equilibrium_positions

SimpleStatics.force_vector

SimpleStatics.global_stiffness_matrix

SimpleStatics.joint_ids

SimpleStatics.member_angle

SimpleStatics.member_ids

SimpleStatics.member_stiffness_matrix

SimpleStatics.n_dofs

SimpleStatics.n_joints

SimpleStatics.n_members

SimpleStatics.original_array

SimpleStatics.plot_setup

SimpleStatics.set_force!

SimpleStatics.solve_displacements

SimpleStatics.solve_member_stresses

SimpleStatics.solve_reaction_forces

SimpleStatics.terminal_joints

SimpleStatics.AnchorConstraint — Type

Completely restrictive constraint that restricts all movement.

SimpleStatics.NoConstraint — Type

Default constraint that restricts no movement.

SimpleStatics.UnorderedPair — Type

Create a pair of items whose order is independent in hashes, equality, and comparisons.

SimpleStatics.XRollerConstraint — Type

Restricts movement along the vertical axis only, permitting horizontal movement.

SimpleStatics.YRollerConstraint — Type

Restricts movement along the vertical axis only, permitting horizontal movement.

SimpleStatics.add_joint! — Method

Add a joint to the setup and return the index referring to that joint. The position should be set in meters.

SimpleStatics.add_member! — Method

Add a member to the setup and return the index referring to that member.

SimpleStatics.constrained_array — Method

Construct a view of an array where constrained indices, indicated by a mapping, are removed. This is useful for solving matrices as zero rows will cause det(M)==0.

SimpleStatics.constrained_dof_array_mapping — Method

Find a mapping m[i] -> j where i is the index of a constrained dof array and j is the corresponding index of the unconstrained array.

SimpleStatics.dof_indices — Method

Get the indices of the x and y rows, respectively, in a square DOF-like matrix (e.g., global stiffness matrix) for a vertex (joint) index.

SimpleStatics.equilibrium_positions — Method

equilibrium_positions(setup, displacements)

Given a setup and its displacements, calculate the equilibrium positions of all of the joints. Returned indices match joint indices.

SimpleStatics.force_vector — Method

Get a vector of the unconstrained forces in the setup.

SimpleStatics.global_stiffness_matrix — Method

Get the stiffness matrix of the entire setup with respect to the global coordinate system.

SimpleStatics.joint_ids — Method

Get an iterator which traverses through all joint ids.

SimpleStatics.member_angle — Method

Get the angle, in radians, the member makes with the global coordinate system's positive X axis.

SimpleStatics.member_ids — Method

Get an iterator which traverses through all member ids.

SimpleStatics.member_stiffness_matrix — Method

Get the global stiffness matrix for a StaticSetup in Newtons / Meter.

SimpleStatics.n_dofs — Method

Naive number of dergees of freedom in the system (i.e., before constraints).

SimpleStatics.n_joints — Method

Get the total number of joints.

SimpleStatics.n_members — Method

Get the total number of members.

SimpleStatics.original_array — Method

Recover the original array where DOF indices will match with the StaticSetup.

SimpleStatics.plot_setup — Function

plot_setup(setup, name="default"; [dsize=800, padding, displacements, stresses, reactions])

Arguments

setup::StaticSetup: The setup to plot.

name::String: The filename or path (without extension) of the image to be created.

dsize::Integer: The diagonal pixel size of the image. The actual width and height will depend on the setup being plotted.

padding::Float64: The amount of extra distance to include on the border of the setup boundaries. 0 padding will make the image as small as possible, and you will likely lose details as they go off of the screen. Default is 0.4.

Additional details that can be included.

displacements: Previously calculated displacements using solve_displacements.

stresses: Previously calculated member stresses using solve_member_stresses.

reactions: Previously calculated reaction forces using solve_reaction_forces.

SimpleStatics.set_force! — Method

Set the force at a joint, in Newtons.

SimpleStatics.solve_displacements — Method

solve_displacements(setup)

Solve the distances that each joint moves to keep the system stationary. The results of this function are necessary for most other properties that can be calculated.

SimpleStatics.solve_member_stresses — Method

solve_member_stresses(setup, displacements)

Using already calculated displacements, determine the stresses each member undergoes when at equilibrium. Compressive stresses are negative and tensile stresses are *positive, indexes match member ids.

SimpleStatics.solve_reaction_forces — Method

solve_reaction_forces(setup, displacements)

Using already calculated displacements, find the forces that each constrained node needs to keep the system stationary.

SimpleStatics.terminal_joints — Method

Get the joints IDs that a member connects, found by its member ID.